League Fİxture Creation Program
This project was created with the aim to create an efficient and systematic fixture  system for sports leagues and to resolve match scheduling problems in sports organizations. The aim is to design fair, conflict-free schedules, taking into account constraints such as team availability, venue constraints, and balance of home-field. There are as many home as away matches, making it fair, and preventing schedule clashes.  Using the Round Robin algorithm to plan a balanced competition works well for various sports leagues and tournaments. The system dynamically manages external factors such as space constraints and team availability to optimize scheduling.  On the technical side of the project, it has a powerful backend built into Laravel, a simple and functional front-end for optimally managing data, and PHP with MySQL. HTML, CSS and modern Tailwind CSS are used for visual design.  Thus, this project produces an optimized fixture creation solution, which sports organizations some automated energy usage for efficient match scheduling with technological support.